Create Psychometric_properties
Function that can be used to calculate the within-person and between-person Cronbach's alpha for a construct.

To use the function, you need to enter a dataset which only consists of columns for the seperate items as well as a column with the participant numbers. Additionally, you need to specify the name of the column with the participant numbers with the participant statement.